Refactor testing code.
