+++ test/in_memory_async_transporter.h
@@ -1,0 +1,34 @@
+#pragma once
+
+#include <collector.pb.h>
+#include <lightstep/transporter.h>
+#include <exception>
+#include <iostream>
+#include <mutex>
+#include <vector>
+
+namespace lightstep {
+class InMemoryAsyncTransporter : public AsyncTransporter {
+ public:
+  void Send(const google::protobuf::Message& request,
+            google::protobuf::Message& response,
+            AsyncTransporter::Callback& callback) override;
+
+  void Write();
+
+  void Fail(std::error_code error);
+
+  const std::vector<collector::ReportRequest>& reports() const {
+    return reports_;
+  }
+
+  const std::vector<collector::Span>& spans() const { return spans_; }
+
+ private:
+  const google::protobuf::Message* active_request_;
+  google::protobuf::Message* active_response_;
+  AsyncTransporter::Callback* active_callback_;
+  std::vector<collector::ReportRequest> reports_;
+  std::vector<collector::Span> spans_;
+};
+}  // namespace lightstep
